Revert "Apply Builder Design Pattern to the API Integrator"
This reverts commit e930c86653503bcce28ee7532fa9c886fed1ddd8.
